+++ R/BASiCS-simulate.R
@@ -10,8 +10,8 @@
 #' @details
 #' This function is just a wrapper around \code{\link[BASiCS]{BASiCS_Sim}} that
 #' takes a \code{\link{BASiCSParams}}, runs the simulation then converts the
-#' output to an \code{\link[scater]{SCESet}} object. See
-#' \code{\link[BASiCS]{BASiCS_Sim}} for more details of how the simulation
+#' output to a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.
+#' See \code{\link[BASiCS]{BASiCS_Sim}} for more details of how the simulation
 #' works.
 #'
 #' @return SingleCellExperiment containing simulated counts
@@ -131,7 +131,7 @@ BASiCSSimulate <- function(params = newBASiCSParams(), verbose = TRUE, ...) {
     sim <- SingleCellExperiment(assays = list(counts = counts),
                                 rowData = features,
                                 colData = cells,
-                                metadata = list(params = params))
+                                metadata = list(Params = params))
 
     if (verbose) {message("Done!")}
     return(sim)