+++ R/scDD-estimate.R
@@ -10,9 +10,9 @@
 #'        interest.
 #' @param params SCDDParams object to store estimated values in.
 #' @param verbose logical. Whether to show progress messages.
-#' @param BPPARAM A \code{\link[BiocParallel]{BiocParallelParam}} instance
-#'        giving the parallel back-end to be used. Default is
-#'        \code{\link[BiocParallel]{SerialParam}} which uses a single core.
+#' @param BPPARAM A \code{\link{BiocParallelParam}} instance giving the parallel
+#'        back-end to be used. Default is \code{\link{SerialParam}} which uses a
+#'        single core.
 #' @param ... further arguments passed to or from other methods.
 #'
 #' @details
@@ -25,9 +25,12 @@
 #'
 #' @examples
 #' \dontrun{
+#' # Load example data
+#' library(scater)
 #' data("sc_example_counts")
+#'
 #' conditions <- sample(1:2, ncol(sc_example_counts), replace = TRUE)
-#' params <- scDDEstimate(sc_example_counts, conditions)
+#' params <- scDDEstimate(sc_example_counts, conditions = conditions)
 #' params
 #' }
 #' @importFrom BiocParallel SerialParam
@@ -58,7 +61,8 @@ scDDEstimate.matrix <- function(counts, params = newSCDDParams(),
 
     counts <- SingleCellExperiment(assays = list(counts = counts),
                                    colData = data.frame(condition = conditions))
-    scDDEstimate.default(counts, "condition", params, verbose, BPPARAM)
+    scDDEstimate.default(counts, params, verbose, BPPARAM,
+                         condition = "condition")
 }
 
 #' @rdname scDDEstimate
@@ -68,7 +72,8 @@ scDDEstimate.SingleCellExperiment <- function(counts,
                                               verbose = TRUE,
                                               BPPARAM = SerialParam(),
                                               condition = "condition", ...) {
-    scDDEstimate.default(counts, condition, params, verbose, BPPARAM)
+    scDDEstimate.default(counts, params, verbose, BPPARAM,
+                         condition = condition)
 }
 
 #' @rdname scDDEstimate