+++ R/splat-simulate.R
@@ -37,10 +37,10 @@
 #' The final output is a
 #' \code{\link[SingleCellExperiment]{SingleCellExperiment}} object that
 #' contains the simulated counts but also the values for various intermediate
-#' steps. These are stored in the \code{\link[SummarizedExperiment]{colData}}
-#' (for cell specific information), \code{\link[SummarizedExperiment]{rowData}}
-#' (for gene specific information) or \code{\link[SummarizedExperiment]{assays}}
-#' (for gene by cell matrices) slots. This additional information includes:
+#' steps. These are stored in the \code{\link{colData}} (for cell specific
+#' information), \code{\link{rowData}} (for gene specific information) or
+#' \code{\link{assays}} (for gene by cell matrices) slots. This additional
+#' information includes:
 #' \describe{
 #'     \item{\code{phenoData}}{
 #'         \describe{
@@ -173,7 +173,7 @@ splatSimulate <- function(params = newSplatParams(),
     features <- data.frame(Gene = gene.names)
     rownames(features) <- gene.names
     sim <- SingleCellExperiment(rowData = features, colData = cells,
-                                metadata = list(params = params))
+                                metadata = list(Params = params))
 
     # Make batches vector which is the index of param$batchCells repeated
     # params$batchCells[index] times
@@ -249,7 +249,9 @@ splatSimulatePaths <- function(params = newSplatParams(), verbose = TRUE, ...) {
 
 #' Simulate library sizes
 #'
-#' Simulate expected library sizes from a log-normal distribution
+#' Simulate expected library sizes. Typically a log-normal distribution is used
+#' but there is also the option to use a normal distribution. In this case any
+#' negative values are set to half the minimum non-zero value.
 #'
 #' @param sim SingleCellExperiment to add library size to.
 #' @param params SplatParams object with simulation parameters.
@@ -257,14 +259,22 @@ splatSimulatePaths <- function(params = newSplatParams(), verbose = TRUE, ...) {
 #' @return SingleCellExperiment with simulated library sizes.
 #'
 #' @importFrom SummarizedExperiment colData colData<-
-#' @importFrom stats rlnorm
+#' @importFrom stats rlnorm rnorm
 splatSimLibSizes <- function(sim, params) {
 
     nCells <- getParam(params, "nCells")
     lib.loc <- getParam(params, "lib.loc")
     lib.scale <- getParam(params, "lib.scale")
+    lib.norm <- getParam(params, "lib.norm")
 
-    exp.lib.sizes <- rlnorm(nCells, lib.loc, lib.scale)
+    if (lib.norm) {
+        exp.lib.sizes <- rnorm(nCells, lib.loc, lib.scale)
+        min.lib <- min(exp.lib.sizes[exp.lib.sizes > 0])
+        exp.lib.sizes[exp.lib.sizes < 0] <- min.lib / 2
+    } else {
+        exp.lib.sizes <- rlnorm(nCells, lib.loc, lib.scale)
+    }
+
     colData(sim)$ExpLibSize <- exp.lib.sizes
 
     return(sim)
@@ -351,7 +361,7 @@ splatSimBatchEffects <- function(sim, params) {
 #'
 #' @return SingleCellExperiment with simulated batch means.
 #'
-#' @importFrom SummarizedExperiment rowData rowData<-
+#' @importFrom SummarizedExperiment rowData rowData<- colData
 splatSimBatchCellMeans <- function(sim, params) {
 
     nBatches <- getParam(params, "nBatches")
@@ -697,22 +707,73 @@ splatSimTrueCounts <- function(sim, params) {
 #' @importFrom stats rbinom
 splatSimDropout <- function(sim, params) {
 
-    dropout.present <- getParam(params, "dropout.present")
+    dropout.type <- getParam(params, "dropout.type")
     true.counts <- assays(sim)$TrueCounts
+    dropout.mid <- getParam(params, "dropout.mid")
+    dropout.shape <- getParam(params, "dropout.shape")
+    cell.names <- colData(sim)$Cell
+    gene.names <- rowData(sim)$Gene
+    nCells <- getParam(params, "nCells")
+    nGenes <- getParam(params, "nGenes")
+    nBatches <- getParam(params, "nBatches")
+    nGroups <- getParam(params, "nGroups")
+    cell.means <- assays(sim)$CellMeans
 
-    if (dropout.present) {
-        cell.names <- colData(sim)$Cell
-        gene.names <- rowData(sim)$Gene
-        nCells <- getParam(params, "nCells")
-        nGenes <- getParam(params, "nGenes")
-        dropout.mid <- getParam(params, "dropout.mid")
-        dropout.shape <- getParam(params, "dropout.shape")
-        cell.means <- assays(sim)$CellMeans
+    switch(dropout.type,
+           experiment = {
+               if ((length(dropout.mid) != 1) || length(dropout.shape) != 1) {
+                   stop("dropout.type is set to 'experiment' but dropout.mid ",
+                        "and dropout.shape aren't length 1")
+               }
+
+               dropout.mid <- rep(dropout.mid, nCells)
+               dropout.shape <- rep(dropout.shape, nCells)
+           },
+           batch = {
+               if ((length(dropout.mid) != nBatches) ||
+                   length(dropout.shape) != nBatches) {
+                   stop("dropout.type is set to 'batch' but dropout.mid ",
+                        "and dropout.shape aren't length equal to nBatches ",
+                        "(", nBatches, ")")
+               }
+
+               batches <- as.numeric(factor(colData(sim)$Batch))
+               dropout.mid <- dropout.mid[batches]
+               dropout.shape <- dropout.shape[batches]
+           },
+           group = {
+               if ((length(dropout.mid) != nGroups) ||
+                   length(dropout.shape) != nGroups) {
+                   stop("dropout.type is set to 'group' but dropout.mid ",
+                        "and dropout.shape aren't length equal to nGroups ",
+                        "(", nGroups, ")")
+               }
+
+               if ("Group" %in% colnames(colData(sim))) {
+                   groups <- as.numeric(factor(colData(sim)$Group))
+               } else {
+                   stop("dropout.type is set to 'group' but groups have not ",
+                        "been simulated")
+               }
+
+               dropout.mid <- dropout.mid[groups]
+               dropout.shape <- dropout.shape[groups]
+           },
+           cell = {
+               if ((length(dropout.mid) != nCells) ||
+                   length(dropout.shape) != nCells) {
+                   stop("dropout.type is set to 'cell' but dropout.mid ",
+                        "and dropout.shape aren't length equal to nCells ",
+                        "(", nCells, ")")
+               }
+           })
+
+    if (dropout.type != "none") {
 
         # Generate probabilites based on expression
         drop.prob <- sapply(seq_len(nCells), function(idx) {
             eta <- log(cell.means[, idx])
-            return(logistic(eta, x0 = dropout.mid, k = dropout.shape))
+            return(logistic(eta, x0 = dropout.mid[idx], k = dropout.shape[idx]))
         })
 
         # Decide which counts to keep